+++ version/doc.go
@@ -9,6 +9,11 @@
 // provide information about their current version and compare it
 // to others. It follows the idea of semantic versioning (see
 // http://semver.org/).
+//
+// Version instances can be created via New() with explicit passed
+// field values or via Parse() and a passed sting. Beside accessing
+// the individual fields two versions can be compared with Compare()
+// and Less().
 package version
 
 // EOF