+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added *Parse()* and *Compare()* to *version*
 - Added package *library* for functions representing the whole
   Go Library; currently only containing *Version()*
+- Extended *etc* to retrieve values out of environment variables
 
 ## 2016-11-01